+++ code/steps/header/model/wtg_models/wtg_model.h
@@ -17,7 +17,7 @@ class WTG_MODEL : public MODEL
         double get_mbase_in_MVA() const;
         double get_one_over_mbase_in_one_over_MVA() const;
 
-        unsigned int get_number_of_lumped_wt_generators() const;
+        double get_number_of_lumped_wt_generators() const;
         double get_rated_power_per_wt_generator_in_MW() const;
         double get_terminal_voltage_in_pu() const;
         complex<double> get_terminal_complex_voltage_in_pu() const;
